VIRO-3745: fix cloud anchors after latest changes
The last change to anchors broke cloud host and resolve
because ARNode.getAnchor() was crashing. Fix this by
setting the anchor in the application thread. After doing this,
it no longer makes sense to have a separate method
in VROARSessionARCore for adding manual anchors, so
remove this method and move the logic to the two classes
that used it. This makes the logic easier to follow at the
expense of some code duplication.

This commit also adds cloud anchor host and resolve
tests, and cleans up the ViroAnchorTests in general.