VIRO-3708: transparent sections are slightly rendered
This was an issue with how we handle partial tone-
mapping. If we had a square geometry with an alpha
texture, the transparent parts of the image would
take their tone-mapping queue from the geometry,
yet the geometry is not what's being rendered there
-- the camera background is.

Fix this with a temporary pre-release hack that solves
against this worst-case problem, and added a story
to create a better solution.