VIRO-3746: first install of ARCore on Pixel2 broken
When installing ARCore on the Pixel 2, ARCore
reports that the device is not compatible. It then
works on the next run. (This diagnosis comes
from looking at the intenral arcore logging).

This appears to be an internal bug to ARCore.
By introducing a (hacky) Thread.sleep of 100
ms, ARCore seems to settle itself and the error
about the Pixel 2 being unsupported goes
away.