VIRO-3704: image textures were overwriting ref materials
Image textures were overwriting the reference
materials in the React MaterialManager's map.
This was because we were using the Surface.java
function setImageTexture, which does not perform
a copy before changing the material.

This method should be removed entirely, but for
now we just make sure it does the copy first. Also
applied this change to Quad.