Fix tone-mapping blending on iOS devices
The previos method, using an explicit branching statement
for writing to the second render buffer, was not working on
iOS. Revert to the previous method using blending. This
method, it turns out, works correctly for all backgrounds
so long as normal blending is activated. The comments to
this commit work thorugh multiple examples.