VIRO-3755: async upload textures to GPU
Prevent renderer lag by asynchronously uploaded
textures to the GPU in the FBX, GLTF, and OBJ loaders.
Further, ensure that the model is not rendered until:

1. The textures are uploded and,
2. The model callback has been invoked.

This prevents all flicker when loading models.

To implement, required creating a new bool in
VRONode that allows the renderer to “hold” the
rendering of a node and all of its children. We also
had to implement a system whereby we can kick
off the async hydration (uploading to GPU) process
directly from the model loaders.