VIRO-3873: hydrateAsync crash on launch
This was happening because the VRODriver doesn’t
get the frame scheduler until the rendering thread
is initialized, but it’s possible for a 3D model to be
loaded prior to rendering init.

This commit fixes this by making the driver *own* the
scheudler, so that it has it on initialization. It is ok if
tasks get added to the scheduler before the renderer
initializes; those tasks will just get handled in the first
few frames.