VIRO-3879 [iOS] DragType FixedToWorld is stuttery
A previous change to properly position a dragged node changed the
dragging to call "setWorldTransform" vs "setPosition" which resulted
in the movement to not be animated. The solution is to add an `animated`
flag to setWorldTransform that's defaulted to false, but can be set to
true in this case. This is not advised for most use cases though