VIRO-3874: instability caused by GL calls off context
The VRODispatchAsyncRenderer call did not guarantee
an EAGL context was bound, but was being used to
perform GL operations. This would result in a loss of sync
between the GPU and CPU, creating severe issues like
buffer overruns during drawing commands.

For example: if we created a vertex array (VAO) without
a valid context for a Geometry, the GPU would not know
that VAO ID was taken. When the context was re-bound,
the GPU would therefore re-assign that VAO, unbeknownst
to the Geometry that was already using that VAO. That
geometry would continue rendering using that VAO, but
using the wrong number of index count. Since the VAO
has changed underneath the Geometry, this would lead
to a buffer overrun during glDrawElements, and a resultant
crash.

The solution is to bind the EAGLContext whenever
VRORendererDispatchAsync is invoked. This is not needed
on Android because our dispatch async on Android is
guaranteed to be invoked with a valid context.