VIRO-3926 ARDragTest - White box clips into/below grey plane
I made this change I think to fix the same issue that commit
f0dcc0c fixed, but mines was problematic. Undoing my change
back then fixes that issue. The commit that made this change
was 215516e which added the FixedToPlane drag type, so we don't
want to revert all of it.