VIRO-2648: create camera image API
Create an API so that developers can process each
image received from the AR camera. This new API
works through a listener, the CameraImageListener.
When installed, the user will get per-frame updates
with the latest camera image, in the form of a
RGBA8888 ByteBuffer.

Tested adding the listener, removing the listener, and
displaying the listener content as bitmaps.

Remaining work: test for memory leaks and ensure
no strong reference cycles were created.