VIRO-2648: memory leak fix for camera listener
Fix the last CR by moving the camera image frame
listener to the view instead of the scene, which makes
more sense. Now, the same listener will continue to
operate even if the scene is changed. This refactor
also corrects a previous memory leak where the
listener would not be destroyed when the ViroView
was removed.