VIRO-3195: glyph atlases for VROText
This change moves from having one material and
texture per glyph to using glyph atlases. Now, all
glyphs are rendered to atlases, and all glyphs that
share the same atlas are rendered in the same draw
call. This means most text strings are now rendered
with a single draw call, resulting in significant
performance benefits.

To make this change, VROGlyphAtlas was introduced.
The VROTypeface now attempts to write each glyph
to an atlas, collecting a vector of these atlases. The
VROText then builds the geometry based on the atlases:
one material per atlas.