VIRO-3808: software linear rendering with HDR off broken
When using software gamma correction (software linear
rendering), if HDR was disabled then we would load textures
into linear space even though we were not gamma correcting
on the framebuffer (because we do not use a tone-mapping
pass if HDR is disabled, and gamma correction is embedded
in the tone-mapping shader).

To fix this, create an easier to use function in the VRODriver
that indicates whether linear rendering is actually enabled.
It now correctly returns true if we have a hardware sRGB buffer or
if the tone-mapping+gamma-correction pass is installed.