VIRO-3815: screenshots black on some devices
Screenshots could end up black on certain Android
devices because we were relying on underlying EGL
framebuffer behavior when invoking glReadPixels to
get the screenshot. IF the underlying framebuffer was
multisampled, for example, glReadPixels would fail.

Change this so that we instead read from the
RTT delegate texture, like we do in iOS. This texture is
of a known format so we can correctly set it up for
glReadPixels.

Note: if the RTT texture is floating point, we first
render it to an LDR render target and invoke
glReadPixels on the LDR target.