VIRO-3987: Viro3DObject Animation Does not Restart
3D Models with inflated nodes were not correctly pausing their animations
when told to, resulting in attempts to replay expired transactions that then
fails an animation restart (animation state gets stuck in a non-replayale state).

The root cause of this was because we were calling the native VRONode.addChildNode
twice when inflating a 3D model's subtree, for the same node. Thus when playing
the animation, the duplicated node also triggers callback state changes onto the
single Executable Animation object, resulting in improper states that can potentially
cause the issue above.