VIRO-3998 [Paramount| Android] Crash when exiting AR
- we were updating the ARCore config on the background thread and we
were seeing this crash upon exiting AR quickly, so it must be that we
were deleting references in the renderer thread, but then still working
on the background thread. I just moved the updating of the ARCore config
to the renderer thread. I tried to repro, and wasn't able to do so (it's
not a 100% repro anyways)