VIRO-1902 [3D text] vectorize glyphs
Add the logic to vectorize glyphs instead of
always rendering them to a bitmap. This generates
the contour of the font outline from FreeType,
then triangulates it using poly2tri. The triangles
are then piped into VROText3D where they are used
to create a VROGeometry.

Currently only one Material is supported but this
can be easily extended to support different materials
for different parts of the text (e.g. front, back, sides).