VIRO-1902 [3D Text] ViroCore API
ViroCore API for 3D text. When building this API,
realized that instead of having an entirely separate
VROText3D object, it’s better to have a single VROText
object, which chooses how to construct its geometry
via the extrusion property. If extrusion is 0, we create
a Bitmap text, otherwise we create a 3D (Vectorized)
text.

Added logic to preserve the materials set on the Text
between updates. Previously, whenever a Text was
reconstructed, its materials would be overwritten. Now,
the previously installed materials are modified. This
is more important now than before because 3D text
is far more likely to use material properties for lighting
and other effects.