VIRO-4052 [material instability] FBX holding weak refs
FBX loader was holding onto weak references
to material visuals, instead of to the material
itself. This could result in the parent material
disappearing and deleting its underlying visual,
creating a crash if the callback then attempted
to access the visual.

Fix this by instead holding onto a weak pointer
of the parent material, so that we can check if
the material disappeared before accessing the
visual.