VIRO-3582 [transparent occlusion] set render order
The second step to getting transparent occlusion
to work is the ability to set the render order of materials.
This was already possible in ViroRenderer but needed
to be added to ViroCore. Additionally, we had to make
the render order accept negative values.

This is necessary to ensure that the occluding surface
is rendered before the objects to be occluded.