VIRO-4151 [wasm build] VROAtomic template
Webassembly does not support atomic types.
Previously, we got around this by using #if
directives everywhere where we used atomic.
However, recent changes to Viro make heavier
use of atomic variables so this was no longer
tenable.

This commit instead introduced the VROAtomic
template. For all platforms but Wasm this
is simply an alias for std::atomic. For Wasm
this is instead just the data type itself.

E.g. for iOS, VROAtomic<int> is std::atomic<int>,
but for Wasm is just int.

This breaks, however, since we use load() and
store() operations on std::atomic, which are not
available on primitive data types. It turns out
we didn’t need to use these: instead we can
replace load() and store() with the assignment
operator, as that automatically calls load and
store if needed.

This commit therefore removes all uses of
load() and store().