[Layered animations] Enable using bone local transforms
For layered animation support, it’s easier to use
local bone transformations instead of concatenated
(head to toe) bone transformations. This commit
works with the latest changes in the FBX loader
to bring the bone local transformations in, and
expands the VROSkinner so that it can handle three
types of bone transformation. These three types
are described in VROBone.h and VROSkinner.h.