Convert VROSkinner to shared_ptr
Previously VROSkinner was a unique_ptr. This worked
fine but there have a been a few cases now where it
made sense to tie the skinner to the skeletal animations
to which it’s associated as well, and the unique_ptr makes
those transitions difficult.

(Note this commit does *not* tie skeletal animations to
the skinner, it only switches from unique to shared)