Layered skeletal animations
Add support for layered skeletal animations. These
are useful when we have multiple animations running
on the same skeleton, either applying to different
parts of the skeleton, or applying with different weights
to the same part of the skeleton.

We accomplish this with the new VROLayeredSkeletonAnimation
API. Given a list of animations defiend by their name,
invidiual weights can be assigned to each bone. The
animation will blend the local bone transformation matrices
and produce a unified animation that can be run in a
VROTransaction.

This required extending the skeletal animation system
to take local bone transformation matrices, and changing
skeletal animations so that they hold their associated
skinner instead of their skeleton.