VIRO-4317: blend modes corrupt blitting
If the last item rendered had a blend mode of
multiply, subtract, or screen, that would cause
blitting operations to fail, resulting in a black
screen.

Fixed this by making blit operations set the
blend mode to the default, Alpha. This commit
also consolidates the setting of all blit operations
in a single place, VRORenderUtil, to reduce
code duplication.