VIRO-4098: polyline thickness distortion
Polyline thickness was distorted at 90 degree
angles because we were always creating the
line stroke along the Y axis (e.g., the thickness
of the line ran up and down the Y axis). This
commit fixes this to make polylines effectively
billboard: the axis used when stroking the line
is the cross product of the camera view vector
and the *direction* of the line segment. In other
words, it’s the vector orthogonal to those two
vectors.

This works except poses some problems with
the circular endcaps, which will be addressed
in the next commit for this story.

This commit also fixes issues with appending
points to polylines with no existing points, and
problems with updating the bounding box of
lines when appending points.