VIRO-4302 [iOS] Sounds don't get onFinish callbacks when loop=true
We were setting AVAudioPlayer's numberOfLoops to -1 in order to loop the
audio. Side effect was, AVAudioPlayer didnt trigger the sound finished
callback in that case. Fixed it by not setting it in that class, but instead
handling looping in bridge manually (by seeking to 0 and replaying) after
firing the onFinish call back