VIRO-4098: correct handling of polyline endcaps
Fix a normalization bug with polyline segments,
and correctly billboard the circular polyline endcaps.
This methodology used is described in VROPolyline.h,
in the custom shader code.

Essentially, instead of computing each triangle
in the circle on the CPU, we now compute the triangles
on the GPU. This way we can billboard them with
respect to the current camera view vector.

The endcap circle, in other words, always faces
the camera view vector (we rotate the polyline
direction ray about the camera view vector to render
the triangles that compose the circle).