Enable Dynamic Wrap Mode Modification
Previously, setting the wrap mode on a material after it has been rendered
did not do anything / result in any visual changes (bug). This change fixes
that by updating the substrate dynmaically (if it's hydrated) whenever a
new warp mode is set.

Before, changing the wrap mode of a material after it has been hydrated
did not result in any visual changes (bug). This fix enables to do that
dynamically, by updating t