Fixed Dragging Issues with FixedToPlane
Fixed two main issues that causes dragging to "pop" or "jump" in fixed to plane mode.

Firstly, for ARInputControllers, screenTouchEvents do not necessarily occur before
a hit test calculation that is driven by the renderer. As such, dragging hit test
calculations done on a buttonDown event is un-reliable and incorrect. The fix is to
move the calculation of originalHitLocations for draggable nodes to occur during the
processDragging event, where hit test results are more up to date.

Secondly, fixedToPlane calculations themeslves were incorrect - they were calculating a
drag offset from the hit location from on the object towards the plane. This causes the
object to "pop" / be pushed into the plane, as the dragOffset itself is a diff of
intersection points, not hit points. The fix is to calculate the drag offset in reference
to the last hit location on the plane, rather than the object.