Improper Media Recording When Switching Orientations
Rotating the orientation of the device in between recordings / screenshots (not during)
would result in skewed ratios. For example, taking a screenshot in portrait mode,
then rotating the device, and then taking another screenshot in landscape mode, will yield
in an improper saved landscape mode screenshot (it will be skewed).

The fix is to ensure that the proper widths and heights are passed to the appropriate
viro media recording components, so that the aspect ratio is correctly updated.