VIRO-4491: geometry memory leak
Subclasses of geometry were creating two
native-refs: one for the subclass and one for
the parent Geometry object. This fixes the issue
by following the pattern we use in Node.java for
inheritance:

1. The mNativeRef is held by the parent (Geometry)
2. The mNativeRef is disposed by the parent (Geometry)
3. The mNativeRef is *created* by the child (Sphere, Surface, etc.)

To do this, the subclasses tell the parent not to create
its own mNativeRef by invoking the dummy-arg
constructor (same pattern we follow for Node). That
way custom Geometry can create mNativeRef by
invoking the normal constructor.