VIRO-4493: GL deletes on wrong thread on Android
Graphics memory was leaking because we were
not guaranteeing that GL deletions would occur
on the rendering thread. Often, destructors called
on the finalizer (or other) threads would invoke
glDeleteBuffers, glDeleteVertexArrays, or similar,
which would be ineffective because the GL context
was not bound.

Fix this by ensuring all GL deletes occur on the
rendering thread. We do this by holding a series
of thread-safe queues in the VRODriver responsible
for allocating those GL objects. Instead of directly
deleting GL objects, instead we call into the driver,
which queues the object for deletion on the next
frame.

Note that there are still some direct glDelete calls
in the code, in places where we have guaranteed
that we’re on the rendering thread.