VIRO-4254: brightness issues when recording
When recording (using RTT) the brightness was off.
This is because we were using an HDR texture to store
the final rendered scene instead of an LDR texture.
The final blit to the display appears to require an LDR
texture as the source.

To fix this introduced a dedicated render target for RTT
operations, which is an LDR texture. The final scene is
tone-mapped onto this target, which can be then be
blitted onto the screen. Note that gamma correction is
still handled correctly when using hardware blitting as per
OpenGL ES 3.0 spec.