VIRO-4507: excessive object creation on Android
When the camera image listener is on, we create
thousands of objects each second. Reduce this
in this commit by pooling Runnable objects which
were one of the worst offenders, and by eliminating
an Iterator over frame listeners.

Confirmed that object creation is stable and low
when camera image listener is off. We can still
optimize more when the image listener is on but
those changes would be more destabilizing with
limited gain.