VIRO-4493: in-flight protobuf decompression
The ‘Riker bug’ was happening because we were breaking the
memory bound during decompression of the VRX file.

This commit mitigates this by adding “in-flight” loading into
a protobuf direct from a GZIP stream. Previously we went:

```gzip stream --> uncompressed stream --> protobuf```

Now we can directly do:

```gzip stream --> protobuf```

This is accomplished by using the GzipInputStream, a subclass
of ZeroCopyInputStream. This implementation is not included
in protobuf-lite, so it is now manually included in the project. The
namespace on iOS had to be modified as well to avoid conflicts
with system protobuf files.