VIRO-4529: Viro360Video Keeps Playing after X Button Press
Video textures continue to play even after the user has exited the
AR/VR experience (and the renderer is destroyed) on Android. This
occurs as the Garbage Collector can take awhile to clean up video
player objects in java. The fix is to ensure that the viro jni player
is paused as it is destructed.