VIRO-4493: smarter Object3D dispose
The second part to fixing large model memory crashes
is to make the Object3D dipose method more
comprehensive, so that when loading (or reloading)
large models, we have the ability to clear all the memory.

This commit achieves this in three ways:

1. Previously, behavior when calling loadModel
multiple times was unstable. With this change, if
loadModel is called while another model is already
loading, that new model will wait until the current
model finishes loading (for stability), and then will
replace that model afterward.

2. disposeAll methods. New disposeAll methods
can recurse down the tree and dispose all children
along with their geometries, materials, and textures.
These can be used to easily clear all the memory
used by discarded objects.

3. disposeModel method. This method can be invoked
on an Object3D before loading a new model. It
effectively invokes disposeAll on the model currently
loaded in the Object3D.