VIRO-4469: alpha outlines because of tone-mapping
When a tone-mapped partially transparent object was
rendered atop a non-tone-mapped background, we
would see an outline appear becuase of the 0.2 threshold
we used to determine whether a given fragment was
tone-mapped. To fix this, instead of using a cutoff
threshold of 0.2, instead *always* blend between the
tone-mapped and the not-tone-mapped color, using
the tone_mapped factor (from 0 to 1) as the interpolation
factor.