Clamp tone-mapped mixing value
Since glBlendFunci is not available on iOS OpenGL
ES 3, if we use VROBlendMode::Add when rendering
tone-mapped particles over a not-tone-mapped
background, we can end up with a large value stored
in our tone_mapped texture. Before using this value
to mix our not-tone-mapped color with our tone-mapped
color, we need to clamp it.