VIRO-4400: deformed polylines
The new polyline billboarding system was not
multiplying polyline direction vectors by the
normal matrix — as a result, directions were not
rotating correctly when a polyline Node was
rotated.

Note this does not fix a separate issue when
displaying polylines on top of ARImageTargets;
that has a different (unknown) cause.