VIROO-4400: polyline offsets world-space mixup
The polyline offsets were being computed in
world space but were then being added to
the polyline in model space. Fix this by computing
the world space offset in a Geometry modifier,
then applying it to the world space position in
a Vertex modifier.

This fixes the issues seen with polylines in ARImageMarkers.