VIRO-4588: abandoned Object3D memory leaks
Fix two leaks that occur when abandoning (removing)
a VRONode in the midst of its FBX, OBJ, or GLTF loading.

1. The hydrateAsync callback gets stored deep down in
VROTexture. This callback held a strong reference to
VRONode, creating a reference cycle between VRONode
and VROTexture. If the hydrateAsync callback never
completed, this ref cycle would prevent the node and
texture from being cleaned up.

2. The Java JNI object delegate held a strong reference
to the Object3D. This object delegate, in turn, was
held onto by the finishCallback, which was held onto
by the aformentioned hydrateAsync callback. This
caused a strong reference cycle between VROTexture
and Object3D.

Both of these have been fixed here but some leaks
remain when models are interrupted, though they appear
to be more minor.