VIRO-3434: leaking task queue and instability
Fix memory leaks and instability when aborting a
scene while it’s still loading FBX or OBJ models.

The task queue was occasionally leaking because
VRODispatchAsyncRenderer does *not* guarantee
it will be executed; if the renderer dies, remaining
tasks are simply cleared.

This could cause cases where not every task would
get an opportunity to invoke taskQueue->onTaskComplete.
As a result, the taskQueue would not clear
those tasks. We would end up with a strong
reference cycle between the task queue and its
tasks.

To fix this, make the tasks reference only a weak
ptr to the parent task queue. And, in order to
prevent the task queue from being prematurely
deleted, scope it to the VRONode for which it
is loading assets.

In addition, this commit fixes instability caused
by using raw pointers in the OBJ and FBX loaders:
again, because VRORendererDispatchAsync was
not always being invoked, sometimes these
pointers were never freed. Fix this by migrating
all of these raw pointers to shared pointers.

These changes also fixed the deadlock bug,
VIRO-4052.