VIRO-4599: update libs to use libc++_static
Update the protobuf and OpenCV dependencies
so that they’re compiled for libc++_static. See
the ViroProtobuf and ViroOpenCV projects for
details on this update. In this commit we’re simply
replacing the binaries and making the required
CMake changes.

With this commit, ViroRenderer is now compatible
with libc++_static and NDK r18.