VIRO-4674: Automatic Scaling during Calibration
- Refactored VROBodyTrackerController updateTrackingJoints() to better capture
defined phases where joints are projected, filtered, and cached.

- Added ability to automatically resize model to fit the user's Torso during
calibration of the VROBodyTrackerController.

- Reduced reliance on ARHitTest. First done by modifying the calibration flow
to only perform a 'Window' of hit test around the torso to determine the depth
value for placing model, and as well as the position of a plane. This plane
will then be used for all future hit test (we are no longer using ARHittest).

- Modified debug UI visuals, where UI colored labels represent the output of
the body tracker ML, and green cubes represent the output of the filtered
data fed into the IKRig.