VIRO-4781: Custom IK Constraints
- 3D models within IKRigs can contain extra bone structures in between effectors
that are meant to be "fixed". As such, this often leads to visual defects - for example
when a 'fixed' multi-segemented shoulder warps and snaps due to the additional bones.
This patch implements a fix where we have the ability to lock / freeze all intermediary
bones within a rig to retain the original bone structure, while still enabling joint
effectors to continue processing the FABRIK calculation.

- Fixed an issue with rig constructions containing IKNodes.

- Added ability to tweak VROPencil's brush thickness.