Render CoreML model detections to screen
Use the CoreML pipelined model (Yolo v3 + NMS)
and render the detected objects and their bounding
boxes to the screen using CoreGraphics.

We use CoreGraphics by creating a new “glass view”
which, for debug purposes only, sits atop the
VROViewAR.

The rest of this commit is primarily focused on
getting the coordinate transformations correct
from the ML model to the screen’s viewport.