Vision request on serial queue with CIImage
This commit contains two changes that enable 60 FPS
for the full Yolo v3 model on iPhone XS:

1. We no longer convert explicitly convert
from YCrCb to RGB. We instead use CIImage, which
under the hood appears to make iOS perform the
conversion on the neural engine (or GPU). This is
undocumented behavior.

2. Run the vision requests on a dedicated serial
queue, and ensure that only one such request is
executed at a time. This follows the examples set
in Apple’s sample projects.