Separate inferred joint from tracked joint
The IK system (in VROBodyTrackingController) and the
ML system (in VROBodyTrackeriOS) were sharing the same
data structure, but each system requires different fields.

Refactor so that each uses its own data structure: the
ML system will use VROInferredBodyJoint, and the IK
system will use VROBodyJoint.

Introduce a conversion from one structure to another in
VROBodyTrackerController.

This commit also changes the VROBodyTracker delegate
so that it is able to return more than one of each joint
type.

Finally, to fix compilation issues, changes were made to
the unrelated VROBodyRecognitionTest so that it more
closely matches the VROObjectRecognitionTest.