Fixed rotational issue during calibration
Sometimes, the model would be rotated away from the camera in a weird
direction during calibration, often times leading to a mis-alignment
of joints. This is because the model's root rotation is always set
to 0,0,0, and depending on the AR initialized world rotation of the
scene, the model can then therefore face the wrong direction. The fix
is to always billboard the model's root during calibration. This is okay
as we are also effectively going to project the model's points along
a camera facing plane at the root anyways to grab z depth.