VIRO-4829: Maintain Initial End Effector Rotation in IKRig
During tracking, the FABRIK algorithm continously rotates intermediate
joints to achieve the desired end effector location. However, this also
can often times lead to an undesired rotational result applied on that
effector. The solution here is to simply retain the original local rotation
and re-apply that back at the end of the FABRIK algorithm.