Eliminate idling time between CoreML processing
Previously we waited for ARKit to give us a
new frame, then we processed that frame. This
meant that we could end up in a situation
where:

1. ARKit submits frame
2. Neural engine processes frame
3. [Neural engine takes awhile, so it misses the
    next ARKit frame, and waits 32ms for the next
    one]
4. ARKit submits frame
5. Neural engine processes frame

This commit eliminates potential idling time by
always retaining images submitted by ARKit, and
immediately processing them as soon as the
neural engine completes the previous frame.

This commit also:
1. Fixes an error where images were not retained,
leading to sporadic crashes in the CIImage
constructor.
2. Adds a profling system for timing out different
sections of neural engine processing.