Crop and scale instead of scale to fill
CoreML does not document whether it’s
VNImageCropAndScaleOptionScaleToFill
option preserves aspect ratio. Instead, switch
to VNImageCropAndScaleOptionCenterCrop,
which guarantees the aspect ratio is maintained
and centers the subject more fully in the
picture.

For this to work, the _transform used to convert
the model into viewport coordinates had to be
modified as well, to account for the changes in
scale and translation associated with center
cropping.