Bone distance filter
Introduce a bone distance filter that attempts
to remove joints that are wildly out of place
compared to previous frames.

This required revamping the pose filters so that
they support looking at data frame by frame as
opposed to in a single time series.

This filter had limited success: ultimately, if the
ML model keeps mis-identifying a label, filters
can only do so much.