Front-facing camera integration
Introduce VROTrackingType::Front that repurposes
the old “inertial” AR system. ARKit is not used when
Front tracking is declared. Instead, we direclty use
the new VROAVCaptureController to retrieve camera
frames and pipe them onto a background preview
layer and into the ML subsystem.

Note ML alignment is not fully working with this commit.

Additionally, when using a front-facing camera, the
X dimension must be flipped (mirrored), as this is the
standard UX for front-facing camera. This mirroring
is not yet integrated into the renderer.