Front-facing camera coordinate system + mirroring
Replace the rudimentary mirroring system we had
before with a coordinate system that’s consistent
across both front and back cameras.

We now treat the *device* as the center of the
coordinate system, regardless of what camera we’re
using. This means negative Z is into the screen
for the back-facing camera, and positive Z is into
the screen for the front-facing camera. This creates
a consistent set of axes around the device that’s
easier to conceptualize.

To make this work, the *view vector* is changed to
(0, 0, 1) when using the front-facing camera, which
again is also consistent.

Finally, viewport mirroring is disabled as it’s no longer
necessary. To account for the camera image being
mirrored, CoreML points are flipped at their point
of origin (in VROBodyTrackeriOS).