Support VN ScaleFit transforms
This transform is better for the front-facing
camera, because the CenterCrop transform has
a tendency to end up cropping out the subject
from the screen entirely. This doesn’t happen with
the back-facing camera because that image is
much wider, providing ample padding for the
CenterCrop operation.

ScaleFit will also be used as the baseline from which
we introduce dynamic cropping.