Crop + Pad + Scale custom operation
Introduce a custom crop+pad+scale operation
that enables us to take in a custom cropping
rectangle, then have it scaled and padded with
grey exactly like our ML training images.

This will enable dynamic cropping in a future commit.

Note this does not yet work with the back camera
because it uses the Accelerate framework and this
has not been tested on YCbCr.