Dynamic cropping box
Implement dynamic cropping. The cropping box
shifts each frame to tighten around the detected
body, and this box is used in the subsequent
frame for detection.

Notes:

1. To do this we needed to separate the vision to
image and image to viewport transforms, because
we crop in image space, but our final joints are
returned in viewport space.

2. Fixed a bug where cropping was not performing
the second transform of moving from resized space
to the cropped image space.

3. Added debug functionality to draw the dynamic
cropping box in viewport space.

4. The box is filtered through a OneEuroFilter, which
was expanded to support doubles.

Note: the cropping box is still unstable for unknown
reasons; however, static cropping is now
working accurately.