Functional dynamic cropping with front camera
Fix the jitter in front-face dynamic cropping by
changing from float to int. The use of floating point
would corrupt the crop operation.

Also add protection where we change the lower
bound of the crop box when the ankles are not
found, so that if they appear later they’re more likely
to be detected.