Dynamic crop through region of interest
Instead of cropping the image ourselves before
feeding it to CoreML, use the VNImageRequest’s
regionOfInterest to get CoreML to perform the
crop itself on the GPU. This makes dynamic cropping
work on both front and back-facing cameras.

This commit also expands the cropping box in the
direction of unfound joints, to alleviate problems
where fast movement causes the crop box to get
stuck to a size smaller than the subject.