Simplify VROBodyTrackerController for Posemoji
The VROBodyTrackerController contained a lot
of logic not needed for Posemoji, including IK
rigging, that complicated the control flow. Instead,
move the VROBodyTrackerController into a new
calss, VROBodyIKController, in order to preserve
that functionality, and repurpose and simplify the
VROBodyTrackerController so that it simply handles
the tasks required by Posemoji: calibration and
joint updates.