Do not process same image twice in tracker
Change the VROBodyTracker so that we no longer
process the same image twice. Previously, it was
possible for the vision queue, if it did not receive a
new image from ARKit, to run ML model on the same
image *again*. This was a waste of resources, and
worse, could lead to jitter because the crop box
would be changing frame by frame. (why is not yet
clear, as skeletons should be stable despite a
changing crop box).

Worse, that crop box jitter could result in joints
being seen then not seen, which in combination
with Posemoji reacting to those changes by turning
sound on and off (which is processor heavy) could
in turn cause further CPU starvation, resulting in
no new ARKit frames, and permanent deadlock.