VIRO-5374: MergeAudioTrack API Crashes
The current mergeAudio API actually overwrrites the current audio URL reference
that VRORecorder uses, with a developer provided 'override-audio' file. As such,
when VRORecorder performs cleanup, it also unintentionally deletes the original
'overide-audio' file, while at the same time causing the renderer to leak the
temp audio file it had originaly created for recording (resulting in an accmulation
of unwanted files).

Thus, any lifecycle handling of the 'override-audio' file should NOT be done by
the renderer, as it dilutes the responsibility of who actually is in charge of
the file and it's clean up (it should be the developer), causing undesired behaviors
and in the case of PoseMoji crashes. The fix is to easily allow the merging of a
separate URL without handling it's cleanup.