VIRO-5395: Failure to Process Upload Video on first Install.
- When performing a screen recording to process video, it triggers
a mic permission request flow that changes the recording state
of the recorder under the app - the app would stop the recorder,
but the VRORecorder's internal mechanisms would automatically start
it again, due to the mic permission request flow.

The fix is to enable the recording of the screen without any mic
and thus mitigates the need for any mic requirements. This patch
also then helps us avoid making any behaviorable changes to how
recording works.

- Also fixed an additional permission issue where we were requesting
for camera access unncessairly during a camera mode of PreRecorded
Video.