Re-enable lightweight VROVideoTexture replacing in ios
Turns out re-allocating a new VROAVPlayerDelegate whenever the video texture
replaces the current video isn't a great idea for applications that replaces
videos constantly (Posemoji) - this would result in allocating a ton of objects
unncessairly when we are just simply pointing the player to a new video source.
It also results in crashes.

Thus, the fix is to revert to the lightweight solution we had before, of simply
repointing the AVPlayer output to the new video item. We also check this time during
renderFrame that both the player item and player are ready to play before rendering
onto the video texture.